Move GenericNode, GroupReference to planner package
This is required to seal PlanNode
