+++ core/trino-main/src/main/java/io/trino/cost/CostCalculatorWithEstimatedExchanges.java
@@ -17,10 +17,10 @@ package io.trino.cost;
 import com.google.errorprone.annotations.ThreadSafe;
 import com.google.inject.Inject;
 import io.trino.Session;
-import io.trino.sql.planner.iterative.GroupReference;
 import io.trino.sql.planner.iterative.rule.DetermineJoinDistributionType;
 import io.trino.sql.planner.iterative.rule.ReorderJoins;
 import io.trino.sql.planner.plan.AggregationNode;
+import io.trino.sql.planner.plan.GroupReference;
 import io.trino.sql.planner.plan.JoinNode;
 import io.trino.sql.planner.plan.PlanNode;
 import io.trino.sql.planner.plan.PlanVisitor;
