+++ core/trino-main/src/test/java/io/trino/sql/planner/iterative/TestMemo.java
@@ -17,12 +17,12 @@ import com.google.common.collect.ImmutableList;
 import io.trino.cost.PlanCostEstimate;
 import io.trino.cost.PlanNodeStatsEstimate;
 import io.trino.sql.planner.PlanNodeIdAllocator;
-import io.trino.sql.planner.Symbol;
+import io.trino.sql.planner.plan.GenericNode;
+import io.trino.sql.planner.plan.GroupReference;
 import io.trino.sql.planner.plan.PlanNode;
 import io.trino.sql.planner.plan.PlanNodeId;
 import org.junit.jupiter.api.Test;
 
-import java.util.List;
 import java.util.Optional;
 
 import static com.google.common.collect.Iterables.getOnlyElement;
@@ -280,35 +280,5 @@ public class TestMemo
     private GenericNode node(PlanNode... children)
     {
         return node(idAllocator.getNextId(), children);
-    }
-
-    private static class GenericNode
-            extends PlanNode
-    {
-        private final List<PlanNode> sources;
-
-        public GenericNode(PlanNodeId id, List<PlanNode> sources)
-        {
-            super(id);
-            this.sources = ImmutableList.copyOf(sources);
-        }
-
-        @Override
-        public List<PlanNode> getSources()
-        {
-            return sources;
-        }
-
-        @Override
-        public List<Symbol> getOutputSymbols()
-        {
-            return ImmutableList.of();
-        }
-
-        @Override
-        public PlanNode replaceChildren(List<PlanNode> newChildren)
-        {
-            return new GenericNode(getId(), newChildren);
-        }
     }
 }
